+++ core/blas/strmm.cpp
@@ -1,7 +1,7 @@
 /**
- * Copyright (c) Huawei Technologies Co., Ltd. 2025. All rights reserved.
- * This file is a part of the CANN Open Software.
- * Licensed under CANN Open Software License Agreement Version 2.0 (the "License").
+ * Copyright (c) 2025 Huawei Technologies Co., Ltd.
+ * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+ * CANN Open Software License Agreement Version 2.0 (the "License").
  * Please refer to the License for details. You may not use this file except in compliance with the License.
  * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
  * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
@@ -107,29 +107,29 @@ AspbStatus asdBlasStrmm(asdBlasHandle handle, asdBlasSideMode_t side, asdBlasFil
         }
 
         if (sideInt == 0 && uploInt == 0) {
-            realM = m;
-            realN = n;
-            realK = m;
+            realM = static_cast<uint32_t>(m);
+            realN = static_cast<uint32_t>(n);
+            realK = static_cast<uint32_t>(m);
             transLeft = (trans == asdBlasOperation_t::ASDBLAS_OP_N) ? 0 : 1;
             transRight = 0;
         } else if (sideInt == 1 && uploInt == 0) {
-            realM = m;
-            realN = n;
-            realK = n;
+            realM = static_cast<uint32_t>(m);
+            realN = static_cast<uint32_t>(n);
+            realK = static_cast<uint32_t>(n);
             transLeft = 0;
             transRight = (trans == asdBlasOperation_t::ASDBLAS_OP_N) ? 0 : 1;
             lessFlag = caseOneFlag;
         } else if (sideInt == 0 && uploInt == 1) {
-            realM = m;
-            realN = n;
-            realK = m;
+            realM = static_cast<uint32_t>(m);
+            realN = static_cast<uint32_t>(n);
+            realK = static_cast<uint32_t>(m);
             transLeft = (trans == asdBlasOperation_t::ASDBLAS_OP_N) ? 0 : 1;
             transRight = 0;
             lessFlag = caseTwoFlag;
         } else if (sideInt == 1 && uploInt == 1) {
-            realM = m;
-            realN = n;
-            realK = n;
+            realM = static_cast<uint32_t>(m);
+            realN = static_cast<uint32_t>(n);
+            realK = static_cast<uint32_t>(n);
             transLeft = 0;
             transRight = (trans == asdBlasOperation_t::ASDBLAS_OP_N) ? 0 : 1;
             lessFlag = caseThreeFlag;
