+++ ops/filter/convolve/convolve/kernel/matmul_base_impl.h
@@ -1,7 +1,7 @@
 /**
- * Copyright (c) Huawei Technologies Co., Ltd. 2025. All rights reserved.
- * This file is a part of the CANN Open Software.
- * Licensed under CANN Open Software License Agreement Version 2.0 (the "License").
+ * Copyright (c) 2025 Huawei Technologies Co., Ltd.
+ * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+ * CANN Open Software License Agreement Version 2.0 (the "License").
  * Please refer to the License for details. You may not use this file except in compliance with the License.
  * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
  * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
@@ -176,7 +176,7 @@ __aicore__ __attribute__((always_inline)) inline void base_matmul_single_core_de
             // not implement yet
             if (trans_b) {
             } else {
-                offset_b = k_idx * K0 * rightCol + n_idx * N0;
+                offset_b = (int64_t)k_idx * K0 * rightCol + (int64_t)n_idx * N0;
             }
 
             int32_t k_actual = (k_idx == (k_loop - 1)) ? (K - k_idx * K0) : K0;
