+++ ops/filter/convolve/convolve/tiling/convolve_tiling.cpp
@@ -1,8 +1,8 @@
 
 /**
- * Copyright (c) Huawei Technologies Co., Ltd. 2025. All rights reserved.
- * This file is a part of the CANN Open Software.
- * Licensed under CANN Open Software License Agreement Version 2.0 (the "License").
+ * Copyright (c) 2025 Huawei Technologies Co., Ltd.
+ * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+ * CANN Open Software License Agreement Version 2.0 (the "License").
  * Please refer to the License for details. You may not use this file except in compliance with the License.
  * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
  * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
@@ -19,8 +19,8 @@
 namespace AsdSip {
 using namespace Mki;
 
-constexpr uint32_t BASE_COL_BLOCK = 128;
-constexpr uint32_t NUM_TWO = 2;
+constexpr uint64_t BASE_COL_BLOCK = 128;
+constexpr uint64_t NUM_TWO = 2;
 
 AsdSip::AspbStatus ConvolveTiling(const LaunchParam &launchParam, KernelInfo &kernelInfo)
 {
@@ -38,8 +38,8 @@ AsdSip::AspbStatus ConvolveTiling(const LaunchParam &launchParam, KernelInfo &ke
 
     OpParam::Convolve param = AnyCast<OpParam::Convolve>(launchParam.GetParam());
 
-    uint64_t workspaceColNum = static_cast<uint64_t>(BASE_COL_BLOCK + param.kernelLen * NUM_TWO - NUM_TWO);
-    uint64_t workspaceRowNum = static_cast<uint64_t>(workspaceColNum + param.kernelLen * NUM_TWO - NUM_TWO);
+    uint64_t workspaceColNum = BASE_COL_BLOCK + static_cast<uint64_t>(param.kernelLen) * NUM_TWO - NUM_TWO;
+    uint64_t workspaceRowNum = workspaceColNum + static_cast<uint64_t>(param.kernelLen) * NUM_TWO - NUM_TWO;
 
     tilingDataPtr->signalLen = param.signalLen;
     tilingDataPtr->kernelLen = param.kernelLen;
